requirements.txt
# ── Data ──────────────────────────────────────────────────────────────────────
yfinance>=0.2.36
pandas>=2.1.0
numpy>=1.26.0
scipy>=1.11.0
pyarrow>=14.0.0          # Parquet support for local data storage

# ── Machine Learning ──────────────────────────────────────────────────────────
torch>=2.2.0
stable-baselines3>=2.2.1
gymnasium>=0.29.1
sb3-contrib>=2.2.1       # Optional: RecurrentPPO reference

# ── Experiment Tracking ───────────────────────────────────────────────────────
mlflow>=2.11.0

# ── API ───────────────────────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
python-dotenv>=1.0.0     # Load .env file

# ── Dev / Testing ─────────────────────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0            # FastAPI async test client
black>=24.0.0
ruff>=0.3.0
mypy>=1.9.0
